backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
drf_extra_fields==3.7.0
djoser==2.2.2
psycopg2-binary==2.9.9
Pillow==10.1.0
pytest==7.4.3
pytest-django==4.7.0
python-dotenv==1.0.0
django-filter==23.4
drf-yasg==1.21.7
coreapi==2.3.3
gunicorn==21.2.0
haversine==2.8.0
